refactor: 상수들을 lib/constant.js로 이동 및 콘솔 제거
